wechat-bot/requirements.txt
wechaty==1.20.2
wechaty-puppet-wechat==1.18.4
requests==2.31.0
python-dotenv==1.0.0
